+++ docs/source/conf.py
@@ -31,7 +31,7 @@ extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.extlinks',
               'sphinx.ext.intersphinx',
               'sphinx.ext.viewcode',
-              'pulsar.utils.docs.html',
+              #'pulsar.utils.docs.html',
               'pulsarext',
               'redisext']
 
@@ -66,7 +66,6 @@ html_sidebars = {
 }
 exclude_trees = []
 html_additional_pages = {
-    'home': 'home.html',
 }
 
 # -- Options for HTML output ---------------------------------------------------