+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Ver. 0.8.5 - Development
   only when reading the timeout was reset and therefore causing long
   streaming responses to fail.
 * Better handling of streaming responses by the wsgi server
+* Added the :ref:`--reload <setting-reload>` command line option and
+  configuration parameter. If used, pulsar auto-reload code changes (useful
+  during development)
+* Added ``file`` log handler
 
 Ver. 0.8.4 - 2014-Jul-07
 ===========================